Updated pyproject.toml; renamed project
