+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "curry-py"
+name = "curry-fp"
 version = "1.3.1"
 description = "Functional currying in Python. 🥘"
 license = "MIT"
@@ -20,7 +20,7 @@ dependencies = [
 ]
 
 [tool.poetry]
-packages = [{include = "currypy", from = "src"}]
+packages = [{include = "curry_fp", from = "src"}]
 classifiers = [
 ]
