+++ src/curry_fp/__init__.py
