+++ src/curry_fp/curry.py
